# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuthDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessAuthDto"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/recipes": {
      "post": {
        "operationId": "RecipesController_add",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRecipeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ]
      },
      "get": {
        "operationId": "RecipesController_findAll",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GetRecipeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ]
      }
    },
    "/recipes/{id}": {
      "delete": {
        "operationId": "RecipesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Recipes"
        ]
      },
      "put": {
        "operationId": "RecipesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOrUpdateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRecipeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ]
      },
      "get": {
        "operationId": "RecipesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetRecipeDto"
                }
              }
            }
          }
        },
        "tags": [
          "Recipes"
        ]
      }
    },
    "/recipes/{id}/rate": {
      "post": {
        "operationId": "RecipesController_rate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RateRecipeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Recipes"
        ]
      }
    }
  },
  "info": {
    "title": "API",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuthDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The email of the user",
            "example": "test@test.com"
          },
          "password": {
            "type": "string",
            "description": "The password of the user",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SuccessAuthDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "AddOrUpdateRecipeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Recipe Name"
          },
          "description": {
            "type": "string",
            "example": "Recipe Description"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "GetRecipeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123"
          },
          "name": {
            "type": "string",
            "example": "Recipe Name"
          },
          "description": {
            "type": "string",
            "example": "Recipe Description"
          },
          "rating": {
            "type": "number",
            "example": 4.5
          },
          "myRating": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "rating"
        ]
      },
      "RateRecipeDto": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "number",
            "example": 4.5
          }
        },
        "required": [
          "rating"
        ]
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}